requirements.txt
absl-py==1.4.0
google-cloud-error-reporting==1.8.1
Flask==2.2.2
Flask-Cloud-NDB==0.2.0
flask-security==3.0.0
flask-wtf==1.0.1
email-validator==1.3.1
gunicorn==20.1.0
google-cloud-ndb==2.1.0
validators==0.20.0
jqqb-evaluator==0.0.1
requests==2.28.2
opentelemetry-exporter-gcp-trace==1.4.0
opentelemetry-exporter-gcp-monitoring==1.4.0a0
opentelemetry-resourcedetector-gcp==1.4.0a0
opentelemetry-propagator-gcp==1.4.0
opentelemetry-api==1.15.0
opentelemetry-sdk==1.15.0
opentelemetry-instrumentation-requests==0.36b0
opentelemetry-instrumentation-flask==0.36b0
opentelemetry-instrumentation-grpc==0.36b0
